# lobel_store_frontend/docs/openapi.generated.yml
swagger: '2.0'
info:
  title: Lobel Store API
  description: Documentation API pour Lobel Store
  version: v1
basePath: /api
consumes:
- application/json
produces:
- application/json
securityDefinitions:
  Basic:
    type: basic
security:
- Basic: []
paths:
...
  /products/products/filter-options/:
    get:
      operationId: products_products_filter_options
      description: Return complete public filter facets, independently of the current
        page.
      parameters:
      - name: search
        in: query
        description: Nom, catégorie, collection ou SKU.
        required: false
        type: string
      - name: category
        in: query
        description: Identifiant de catégorie.
        required: false
        type: integer
      - name: collection
        in: query
        description: Identifiant ou slug de collection.
        required: false
        type: string
      - name: available
        in: query
        description: true si une variante active est en stock.
        required: false
        type: boolean
      - name: min_price
        in: query
        required: false
        type: number
      - name: max_price
        in: query
        required: false
        type: number
      - name: color
        in: query
        required: false
        type: integer
      - name: size
        in: query
        required: false
        type: integer
      - name: ordering
        in: query
        description: name, price, date_created ou sales_count; préfixe - pour décroissant.
        required: false
        type: string
      - name: page
        in: query
        description: A page number within the paginated result set.
        required: false
        type: integer
      - name: page_size
        in: query
        description: Number of results to return per page.
        required: false
        type: integer
      responses:
        '200':
          description: ''
          schema:
            required:
            - count
            - results
            type: object
            properties:
              count:
                type: integer
              next:
                type: string
                format: uri
                x-nullable: true
              previous:
                type: string
                format: uri
                x-nullable: true
              results:
                type: array
                items:
                  $ref: '#/definitions/ProductDetail'
      tags:
      - products
    parameters: []
...
  /products/products/resolve-variants/:
    post:
      operationId: products_products_resolve_variants
      description: ''
      parameters:
      - name: data
        in: body
        required: true
        schema:
          $ref: '#/definitions/ProductDetail'
      responses:
        '201':
          description: ''
          schema:
            $ref: '#/definitions/ProductDetail'
      tags:
      - products
    parameters: []
...
definitions:
...
  Category:
    required:
    - name
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      name:
        title: Name
        type: string
        maxLength: 100
        minLength: 1
      description:
        title: Description
        type: string
      is_active:
        title: Is active
        type: boolean
      date_created:
        title: Date created
        type: string
        format: date-time
        readOnly: true
...
  ProductDetail:
    required:
    - name
    - price
    type: object
    properties:
      id:
        title: ID
        type: integer
        readOnly: true
      name:
        title: Name
        type: string
        maxLength: 200
        minLength: 1
      category:
        $ref: '#/definitions/Category'
      description:
        title: Description
        type: string
      price:
        title: Price
        type: string
        format: decimal
      sales_count:
        title: Sales count
        type: integer
        maximum: 2147483647
        minimum: 0
      is_active:
        title: Is active
        type: boolean
      date_created:
        title: Date created
        type: string
        format: date-time
        readOnly: true
      collections:
        title: Collections
        type: string
        readOnly: true
      media_files:
        title: Media files
        type: string
        readOnly: true
      variants:
        title: Variants
        type: string
        readOnly: true
      image:
        title: Image
        type: string
        readOnly: true
      video:
        title: Video
        type: string
        readOnly: true
      is_available:
        title: Is available
        type: boolean
        readOnly: true
...